+++ pkg/analyzer/lib/src/dart/ast/utilities.dart
@@ -9574,7 +9574,7 @@ class ToSourceVisitor2 implements AstVisitor<void> {
 
   void _writeOperand(Expression node, Expression operand) {
     if (operand != null) {
-      bool needsParenthesis = operand.precedence < node.precedence;
+      bool needsParenthesis = operand.precedence2 < node.precedence2;
       if (needsParenthesis) {
         sink.write('(');
       }